Split model in core to match uniffi
