+++ src/core/mod.rs
@@ -10,7 +10,6 @@ macro_rules! inner_attr_to_each {
 pub(crate) mod error;
 pub(crate) mod graph;
 pub(crate) mod orchestrator;
-pub(crate) mod pipeline;
 pub(crate) mod store;
 pub(crate) mod util;
 pub(crate) mod validation;
