+++ src/uniffi/model/pod.rs
@@ -2,8 +2,8 @@ use crate::{
     core::{
         crypto::{hash_blob, hash_buffer},
         model::{
-            deserialize_pod, deserialize_pod_job, serialize_hashmap, serialize_hashmap_option,
-            to_yaml,
+            pod::{deserialize_pod, deserialize_pod_job},
+            serialize_hashmap, serialize_hashmap_option, to_yaml,
         },
         util::get,
         validation::validate_packet,
